requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
WTForms==3.2.1
Werkzeug==3.1.5
PyMySQL==1.0.3
python-dotenv==1.0.0
email-validator==2.0.0
cryptography==41.0.0
itsdangerous==2.2.0
gunicorn==21.2.0
